backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
yt-dlp==2024.1.26
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
python-dotenv==1.0.0
